+++ tests/test_censor.py
@@ -1,0 +1,12 @@
+"""
+__author__ = Hagai Hargil
+"""
+
+
+import unittest
+from os import sep
+from pysight.censor_tools import CensorCorrection
+
+
+class TestCensorTools(unittest.TestCase):
+    pass